requirements.txt
opencv-python==4.7.0.72
numpy==1.24.3
pyserial==3.5
qrcode==7.4.2
pillow==9.5.0
pyzbar==0.1.9 